Item sets working
This now successfully reports on the parameters of each item. Next
step is to convert this into a dict in which index 1 equates with
parameter 1, etc, removing the ITEM_COUNTER data.

Also, this does not yet understand the structure of the SCORE file,
so adding some information about the opening short and closing
trailer will (hopefully) make it easier to give the parametric
functions a definitive end point. (At present, the make_item_set
function will blindly convert the trailer binary data.)