Attempting to define loop boundary (not working)
The aim is to use the file's short to define the loop boundary;
ie, for each call of the make_item_set() function, subtract the
total number of floats (item_counter + parameters) from the file's
TOTAL_COUNTER, and only continue the loop whilst the TOTAL_COUNTER
is > 0 (though this may need adjusted to reflect the trailer?).

A second function is being added to save the output of each call of
make_item_set() to a numeric value which increments each time the
make_item_set() function is called. This may be totally unnecessary,
so may be abandoned.